+++ tools/node-depman/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+#### 🚀 Updates
+
+- Updated version detection to loop through all possible fields (`packageManager`, `engines`, etc) to find an applicable version, instead of failing on the first invalid version.
+
 #### 🐞 Fixes
 
 - Fixed an issue where auth headers were not included in load verion requests, which would cause failures for private registries.